sub_081569A0

+++ src/title_screen.c
@@ -85,7 +85,7 @@ void sub_08149F08(struct TitleStruct* arg0) {
     u8 r0 = arg0->unk128;
     arg0->unk8 |= 1;
     r6 = arg0->unk10;
-    r6->unk10 = (r0 * 20) + (r0 / 5 * 12 + 24);
+    (u16)r6->unk10 = (r0 * 20) + (r0 / 5 * 12 + 24);
     r6->unk12 = 0x7f;
 
     if (sub_08155128(r6) == 0) {
@@ -224,7 +224,7 @@ void sub_0814A39C(struct TitleStruct* arg0) {
     u8 i, r7 = arg0->unk128;
     for (i = 0; i < r7; i++) {
         struct Sprite* cur = &arg0->unk10[gUnk_08D6274A[i]];
-        cur->unk10 = (i * 20) + (i / 5 * 12 + 24);
+        (u16)cur->unk10 = (i * 20) + (i / 5 * 12 + 24);
         cur->unk12 = 0x7f;
         sub_0815604C(cur);
     }
